+++ src/game/tests.rs
@@ -393,6 +393,7 @@ mod vintage_cube_expedition_map;
 mod vintage_cube_exploration;
 mod vintage_cube_expressive_iteration;
 mod vintage_cube_fable_of_the_mirror_breaker;
+mod vintage_cube_fabled_passage;
 mod vintage_cube_faerie_mastermind;
 mod vintage_cube_fanatic_of_rhonas;
 mod vintage_cube_fatal_push;